requirements.txt
scapy
pandas
scikit-learn
numpy
PyYAML
